Update check.js
I updated to run git merge-base on the current git branch and the basetheme/master. When you are working in a branch (e.g. develop) this command keeps reverting basetheme commit and version in theme.json to previous basetheme versions and complaining that there is newer version available, even when theme.json points to the most recent basetheme version.